google/chromeec/acpi: Replace GOOG0006 with Intel's analogous device
GOOG0006 is a device that reports if a system is in tablet mode or
laptop mode.  Intel has an analogous device in ACPI (though that
device can also handle other things like buttons for volume and power).
This change gets rid of GOOG0006 (and thus the related need to
maintain a driver for the device on one or more OSes).

Parts of the change were motivated by what I saw on another laptop,
primarily the sequence and values of the events that are sent to
the new device.

BRANCH=eve
BUG=None
TEST=I confirmed that the change works as expected on my eve.  The
Linux driver doesn't yet have proper handling to ignore events
0xCA and 0xCB (it logs a message in the kernel log and a ? character
appears in the console each time the device enters and exits tablet
mode -- the character doesn't seem to appear while using the GUI).

Signed-off-by: delco@chromium.org

Change-Id: Ie913788fde7402b556a18357b89a5e9020acd98d
Reviewed-on: https://chromium-review.googlesource.com/1060385
Tested-by: Matt Delco <delco@google.com>
Trybot-Ready: Matt Delco <delco@google.com>
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Commit-Queue: Matt Delco <delco@google.com>